+++ crates/prover/src/lib.rs
@@ -8,6 +8,7 @@ pub mod components;
 pub mod errors;
 pub mod preprocessed;
 pub mod prover;
+pub mod public_data;
 pub mod relations;
 pub mod verifier;
